# lib/api-zod/src/generated/types/analysesList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Analysis } from "./analysis";

export interface AnalysesList {
  analyses: Analysis[];
  total: number;
}