# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Analysis {
  id: number;
  userId: number;
  telegramId: string;
  username?: string | null;
  firstName?: string | null;
  imageUrl?: string | null;
  analysisText: string;
  diseaseDetected: boolean;
  cropType?: string | null;
  severity?: string | null;
  createdAt: string;
}