# lib/api-zod/src/generated/types/botStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotStats {
  totalUsers: number;
  totalAnalyses: number;
  todayAnalyses: number;
  activeUsers7d: number;
}