# lib/api-zod/src/generated/types/botUsersList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BotUser } from "./botUser";

export interface BotUsersList {
  users: BotUser[];
  total: number;
}