# lib/api-zod/src/generated/types/recentActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecentActivityItem } from "./recentActivityItem";

export interface RecentActivity {
  items: RecentActivityItem[];
}