# lib/api-zod/src/generated/types/recentActivityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecentActivityItem {
  id: number;
  username?: string | null;
  firstName?: string | null;
  cropType?: string | null;
  diseaseDetected: boolean;
  severity?: string | null;
  analysisText: string;
  imageUrl?: string | null;
  createdAt: string;
}