# lib/db/src/schema/analyses.ts
import { pgTable, serial, integer, text, boolean, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { botUsersTable } from "./bot_users";

export const analysesTable = pgTable("analyses", {
  id: serial("id").primaryKey(),
  userId: integer("user_id").notNull().references(() => botUsersTable.id),
  imageUrl: text("image_url"),
  analysisText: text("analysis_text").notNull(),
  diseaseDetected: boolean("disease_detected").notNull().default(false),
  cropType: text("crop_type"),
  severity: text("severity"),
  createdAt: timestamp("created_at").defaultNow().notNull(),
});

export const insertAnalysisSchema = createInsertSchema(analysesTable).omit({ id: true, createdAt: true });
export type InsertAnalysis = z.infer<typeof insertAnalysisSchema>;
export type Analysis = typeof analysesTable.$inferSelect;